+++ tests/ui/pass_user_items.rs
@@ -15,10 +15,12 @@ impl core::fmt::Display for Kind {
 
 impl Kind {
     pub const VARIANTS: &'static [&'static str] = &["a", "b"];
+    pub const NUMBERS: &'static [&'static str] = &["zero", "one"];
 }
 
 fn main() {
     assert_eq!(Kind::VARIANTS, &["a", "b"]);
+    assert_eq!(Kind::NUMBERS, &["zero", "one"]);
     assert_eq!(<Kind as Variants>::VARIANTS, &[Kind::A, Kind::B]);
     assert_eq!(<Kind as Variants>::NUMBERS, &[0, 1]);
     assert_eq!(Kind::A.to_string(), "user");